+++ seiscat/editdb.py
@@ -42,6 +42,70 @@ def _parse_set_arg(arg):
     return key, val
 
 
+def _replicate(config, fields, rows):
+    """
+    Replicate event in database.
+
+    :param config: config object
+    :param fields: list of fields
+    :param rows: list of rows
+    """
+    for row in rows:
+        eventid = row[fields.index('evid')]
+        version = row[fields.index('ver')]
+        replicate_event_in_db(config, eventid, version)
+
+
+def _delete(config, fields, rows, eventid, version, args):
+    """
+    Delete event from database.
+
+    :param config: config object
+    :param fields: list of fields
+    :param rows: list of rows
+    :param eventid: event ID
+    :param version: event version
+    :param args: parsed arguments
+    """
+    if eventid:
+        version = rows[0][fields.index('ver')]
+        if not args.force:
+            _are_you_sure(
+                f'Delete event {eventid} version {version} from database?')
+    elif version and not args.force:
+        _are_you_sure(
+            f'Delete ALL events of version {version} from database?')
+    elif not version and not args.force:
+        _are_you_sure('Delete ALL events from database?')
+    delete_event_from_db(config, eventid, version)
+
+
+def _set(config, fields, rows, key_values, args):
+    """
+    Set key-value pairs in database.
+
+    :param config: config object
+    :param fields: list of fields
+    :param rows: list of rows
+    :param key_values: list of key-value pairs
+    :param args: parsed arguments
+    """
+    if len(rows) == 1 and not args.force:
+        eventid = rows[0][fields.index('evid')]
+        version = rows[0][fields.index('ver')]
+        _are_you_sure(
+            f'Update event {eventid} version {version} in database?')
+    elif not args.force:
+        _are_you_sure(
+            f'Update {len(rows)} events in database?')
+    for row in rows:
+        eventid = row[fields.index('evid')]
+        version = row[fields.index('ver')]
+        key_values = [_parse_set_arg(arg) for arg in args.set]
+        for key, val in key_values:
+            update_event_in_db(config, eventid, version, key, val)
+
+
 def editdb(config):
     """
     Edit database.
@@ -51,32 +115,26 @@ def editdb(config):
     args = config['args']
     eventid = args.eventid[0]
     version = args.version
+    if eventid == 'ALL':
+        eventid = None
     fields, rows = read_fields_and_rows_from_db(config, eventid, version)
     if not rows:
-        if version:
+        if eventid and version:
             msg = f'Event {eventid} version {version} not found in database'
-        else:
+        elif eventid:
             msg = f'Event {eventid} not found in database'
+        else:
+            msg = 'No events found in database'
         err_exit(msg)
-    if len(rows) > 1:
+    if eventid and len(rows) > 1:
         err_exit(
             f'Event {eventid} has {len(rows)} versions, '
             'please specify version with "--version"')
     if args.replicate:
-        replicate_event_in_db(config, eventid)
+        _replicate(config, fields, rows)
     elif args.delete:
-        version = rows[0][fields.index('ver')]
-        if not args.force:
-            _are_you_sure(
-                f'Delete event {eventid} version {version} from database?')
-        delete_event_from_db(config, eventid, version)
+        _delete(config, fields, rows, eventid, version, args)
     elif args.set:
-        version = rows[0][fields.index('ver')]
-        if not args.force:
-            _are_you_sure(
-                f'Update event {eventid} version {version} in database?')
-        key_values = [_parse_set_arg(arg) for arg in args.set]
-        for key, val in key_values:
-            update_event_in_db(config, eventid, version, key, val)
+        _set(config, fields, rows, args.set, args)
     else:
         err_exit('No action specified. See "seiscat editdb -h" for help')